+++ CHANGELOG.md
@@ -1,8 +1,18 @@
 # CHANGELOG docker-yeoman
 
+0.0.7 adding Travis CI build
+
+0.0.6 adding generator-docker for selftest
+
 0.0.6 refactoring
+
 0.0.5 testing generated app files
+
 0.0.4 first test successful
+
 0.0.3 link yo_src generators to ../node_modules
+
 0.0.2 empty generator works
+
 0.0.1 project created
+